% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtres-package.R
\docType{package}
\name{rtres-package}
\alias{rtres}
\alias{rtres-package}
\title{rtres: Interact with the TRES (Trusted Reversible Encryption Service) API}
\description{
This library contains functions to directly use ZorgTTPs TRES API, without the need to understand the REST API.
}
\author{
\strong{Maintainer}: ZorgTTP \email{info@zorgttp.nl} [copyright holder]

}
\keyword{internal}
